Fix loops over progressions near to MAX_VALUE/MIN_VALUE
 #KT-492 Fixed

For Byte, Char and Short ranges, promote the type of the loop parameter to int
to avoid overflows. For Int and Long ranges at the end of the loop over a
progression we now check if the new (incremented) value of the loop parameter
is greater than the old value iff increment > 0

+++ runtime/src/jet/ByteProgressionIterator.java
@@ -17,7 +17,7 @@
 package jet;
 
 class ByteProgressionIterator extends ByteIterator {
-    private byte next;
+    private int next;
     private final byte end;
     private final int increment;
 
@@ -34,8 +34,8 @@ class ByteProgressionIterator extends ByteIterator {
 
     @Override
     public byte nextByte() {
-        byte value = next;
+        int value = next;
         next += increment;
-        return value;
+        return (byte) value;
     }
 }
